refactor(sync): move sync-related data structures to sync_types.h
